Change the base image to amazeeio/solr.
The base govcms image triggers a precreate-core command for Solr, this inserts the correct values in the docker layers as we inherit. When we call the process again in this container build it doesn't recreate the core (as it has been created previously) so it loads the configuration from the inherited containers and doesn't apply our changes.

- Changes the base image ot amazeeio/sor.